Fix stack overflow on Android, and ensure thread safety
Android's main thread has a very limited stack size.  In older versions of
the OS (e.g. Jellybean and Honeycomb), the stack size is so small that
even initializing JavaScriptCore causes a stack overflow.  To address this,
DispatchQueue has been added to enable running heavy JSC operations on
a worker thread and then synchronizing.  This has the added benefit of
ensuring thread safety, as now any operation that seeks to make changes
to the VM will be put into a single-threaded queue and executed
serially.